+++ code/cell.types/src/types/index.ts
@@ -3,6 +3,7 @@ export * from './types.cell.coord';
 export * from './types.cell.get';
 export * from './types.cell';
 export * from './types.client';
+export * from './types.column';
 export * from './types.compile';
 export * from './types.db';
 export * from './types.electron';
@@ -20,6 +21,7 @@ export * from './types.map';
 export * from './types.ns';
 export * from './types.range';
 export * from './types.refs';
+export * from './types.row';
 export * from './types.schema';
 export * from './types.schema.uri';
 export * from './types.schema.url';
